Refactors guest form layouts and routing
Introduces a dedicated `(with-form)` route group for guest authentication pages. This change centralizes common structural elements, such as the back button and form card wrapper, into a new `(with-form)/layout.tsx`.

The `GuestWithFormSubLayout` component is refactored to focus solely on displaying the page title, icon, and status messages (loading/error).

Splits the `forgot-password` page into server and client components to better manage metadata and client-side interactivity. Relocates `RedirectIfAuth` to the login page for more granular control.

Simplifies the root `(guest)` layout by removing unnecessary wrapping elements.
