+++ src/app/(guest)/(with-form)/_parts/guest-with-form-sub-layout.tsx
@@ -1,0 +1,63 @@
+// types
+import type { ReactNode } from 'react'
+// materials
+import Avatar from '@mui/material/Avatar'
+import Box from '@mui/material/Box'
+import Typography from '@mui/material/Typography'
+// components
+import ErrorCenter from '@/components/Statuses/ErrorCenter'
+import LoadingCenter from '@/components/Statuses/LoadingCenter'
+// parts
+
+export default function GuestWithFormSubLayout({
+    children,
+    title,
+    icon,
+    isLoading = false,
+    isError = false,
+    message,
+}: {
+    children: ReactNode
+    icon: ReactNode
+    isLoading?: boolean
+    isError?: boolean
+    message?: string
+    title: string
+}) {
+    return (
+        <>
+            <Box display="flex" gap={3} alignItems="center">
+                <Avatar
+                    sx={{
+                        bgcolor: () => {
+                            if (isLoading) return 'primary.main'
+                            if (isError) return 'error.main'
+                        },
+                    }}>
+                    {icon}
+                </Avatar>
+
+                <Box>
+                    <Typography component="div" variant="body2">
+                        {process.env.NEXT_PUBLIC_APP_NAME}
+                    </Typography>
+
+                    <Typography
+                        component="div"
+                        variant="h5"
+                        lineHeight="normal">
+                        {title}
+                    </Typography>
+                </Box>
+            </Box>
+
+            <LoadingCenter isShow={isLoading} message={message} />
+
+            <ErrorCenter isShow={isError} message={message} />
+
+            <Box display={isLoading ? 'none' : 'block'} mt={2}>
+                {children}
+            </Box>
+        </>
+    )
+}
